`expected_date` is a field that is added in standard in the list view of Sale Orders. If multiple SOs in the list view have multiple lines, this would read thousands of SOL (hitting the PREFETCH_MAX limit multiple tiems). In extreme cases this could also kill the process, as too many Lines would be fetched at ones

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr