@robodoo 2024-11-14
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/187258.png)](https://mergebot.odoo.com/odoo/odoo/pull/187258)