@emi-odoo 2024-11-14
Script used to test performance:

```py
from odoo.tools.profiler import Profiler, PeriodicCollector
import random
product = self.env["product.template"].search([("name", "=", "test_perf")], limit=1)
if not product:
    product = self.env["product.template"].create({"name": "test_perf"})
product_2 = self.env["product.template"].search([("name", "=", "test_perf_2")], limit=1)
if not product_2:
    product_2 = self.env["product.template"].create({"name": "test_perf_2"})
product_3 = self.env["product.template"].search([("name", "=", "test_perf_3")], limit=1)
if not product_3:
    product_3 = self.env["product.template"].create({"name": "test_perf_3"})

customer = self.env["res.partner"].search([("name", "=", "test_perf")], limit=1)
if not customer:
    customer = self.env["res.partner"].create({"name": "test_perf"})
products = product | product_2 | product_3

sale_orders = self.env["sale.order"].create(
    [
        {
            "partner_id": customer.id,
            "order_line": [
                (0, 0, {"product_id": random.choice(products).product_variant_id.id, "product_uom_qty": 1, "price_unit": random.random()*100, "customer_lead": random.random()*50})
                for _ in range(230)
            ],
        }
        for _ in range(400)
    ]
)

# wait quite a bit
self.env.cr.commit()

# in a second moment
customer = self.env["res.partner"].search([("name", "=", "test_perf")], limit=1)
sale_orders = self.env["sale.order"].search([("partner_id", "=", customer.id)])
from odoo.tools.profiler import Profiler, PeriodicCollector


self.env.invalidate_all()
with Profiler(collectors=[PeriodicCollector(0.001)]) as profiler:
    sale_orders[:50]._compute_expected_date()
    self.env.invalidate_all()
    sale_orders[:100]._compute_expected_date()
    self.env.invalidate_all()
    sale_orders[:200]._compute_expected_date()
    self.env.invalidate_all()
    sale_orders._compute_expected_date()
```

Speedscope before: 
[before.json](https://github.com/user-attachments/files/17749425/before.json)
Speedscope after: 
[after.json](https://github.com/user-attachments/files/17749426/after.json)

Table with comparison:

| Rows | Before (s) | After (s) | Time Saved (s) | % Improvement |
|------|------------|-----------|----------------|---------------|
| 50   | 2.25       | 0.39      | 1.86           | 82.67%       |
| 100  | 6.11       | 0.54      | 5.57           | 91.16%       |
| 200  | 20.11      | 1.04      | 19.07          | 94.83%       |
| 400  | 70.00      | 2.06      | 67.94          | 97.06%       |